chore: delete dead Expo UserContext.ts
- Move TUser type into UserProvider.tsx (only surviving consumer)
- Update 3 screen imports: Interactions, InteractionsHmis, ConsentModal
- Remove dependency on @monorepo/react/shared TOrganization type
