--- libs/expo/betterangels/src/lib/providers/user/UserContext.ts
@@ -1,29 +1,0 @@
-import { TOrganization } from '@monorepo/react/shared';
-import { createContext, Dispatch, SetStateAction } from 'react';
-
-export type { TOrganization };
-
-export type TUser = {
-  id: string;
-  username?: string;
-  firstName?: string;
-  lastName?: string;
-  email?: string | null;
-  organizations: TOrganization[];
-  isOutreachAuthorized?: boolean;
-  hasAcceptedTos?: boolean;
-  hasAcceptedPrivacyPolicy?: boolean;
-  isHmisUser?: boolean;
-};
-
-export interface IUserProviderValue {
-  user: TUser | undefined;
-  setUser: Dispatch<SetStateAction<TUser | undefined>>;
-  isLoading: boolean;
-  refetchUser: () => Promise<void>;
-  isHmisUser: boolean | undefined;
-}
-
-const UserContext = createContext<IUserProviderValue | undefined>(undefined);
-
-export default UserContext;
