+++ libs/expo/betterangels/src/lib/screens/InteractionsHmis/index.tsx
@@ -5,7 +5,7 @@ import { ElementType, useCallback, useState } from 'react';
 import { StyleSheet, View } from 'react-native';
 import { HmisNoteType, toNoteFilterHmis } from '../../apollo';
 import useUser from '../../hooks/user/useUser';
-import { TUser } from '../../providers/user/UserContext';
+import { TUser } from '../../providers/user/UserProvider';
 import {
   Header,
   HorizontalContainer,
